add logic like

+++ api/model/likes.go
@@ -1,9 +1,9 @@
 package model
 
 type Likes struct {
-	ID     int `gorm:"primaryKey" json:"id"`
-	SpotID int `gorm:"not null" json:"spot_id"`
-	UserID int `gorm:"not null" json:"user_id"`
+	ID     int64 `gorm:"primaryKey" json:"id"`
+	SpotID int64 `gorm:"not null" json:"spot_id"`
+	UserID int64 `gorm:"not null" json:"user_id"`
 
 	// Relations
 	Spot *Spots `gorm:"foreignKey:SpotID" json:"spot,omitempty"`
